Compatibility fix
mikutap.css (1:1311)	The value for the "width" property is invalid. CSS 2.1 allows this value to be a length, a percentage, or "auto". CSS3 allows this value to be "[<length> | <percentage>] && [border-box | content-box]? | available | min-content | max-content | fit-content | auto". This property specifies the width of the content area or border area (depending on 'box-sizing') of certain boxes.